refactor: Modify directory structure (#4)
